tests/requirements.txt
# Test dependencies for StromaAI
pytest>=8.0
pytest-mock>=3.14
requests-mock>=1.12

# Gateway tests
pytest-asyncio>=0.24
httpx>=0.27.0
respx>=0.21.0
PyJWT[crypto]>=2.9.0
cryptography>=43.0.0

# CLI tests
rich>=13.0.0
